requirements.txt
langgraph
langchain
langchain-community
langchain-google-genai
faiss-cpu
streamlit
pandas
pdfplumber
python-pptx
openpyxl
python-dotenv
sentence-transformers
langchain-huggingface
